trading_app/trading_app/requirements.txt
Flask
Flask-SQLAlchemy
pandas
numpy
yfinance
scikit-learn
plotly